refactor: generate user flow process
