+++ packages/cli/src/lib/commands/init/constants.ts
@@ -1,1 +1,6 @@
+import { FlowExamples } from './types';
+
 export const SETUP_CONFIRM_MESSAGE = 'user-flow CLI is set up now! 🎉';
+export const FlowExampleMap: Record<FlowExamples, string> = {
+  'basic-navigation': 'basic-navigation.ts'
+};
