+++ packages/cli/src/lib/commands/init/index.ts
@@ -1,7 +1,7 @@
 import { YargsCommandObject } from '../../core/yargs/types';
 import { log, logVerbose } from '../../core/loggin';
 import { INIT_OPTIONS } from './options';
-import { addUserFlow, getExamplePathDest, getInitCommandOptionsFromArgv } from './utils';
+import { getExamplePathDest, getInitCommandOptionsFromArgv } from './utils';
 import { collectRcJson } from './processes/collect-rc-json';
 import { askToSkip } from '../../core/prompt';
 import { run } from '../../core/processing/behaviors';
@@ -9,7 +9,7 @@ import { readFile } from '../../core/file';
 import { SETUP_CONFIRM_MESSAGE } from './constants';
 import { RcJson } from '../../types';
 import { updateRcJson } from './processes/update-rc-json';
-import { mkdirSync, readdirSync } from 'fs';
+import { generateUserFlow, userflowIsNotCreated } from './processes/generate-userflow';
 
 export const initCommand: YargsCommandObject = {
   command: 'init',
@@ -21,25 +21,12 @@ export const initCommand: YargsCommandObject = {
       const cfg = getInitCommandOptionsFromArgv(argv);
       logVerbose('Init options: ', cfg);
 
-      const exampleName = 'basic-navigation';
-      const userflowIsNotCreated = (cfg?: RcJson) => Promise.resolve(cfg ? readFile(getExamplePathDest(exampleName, cfg.collect.ufPath)) === '' : false);
-
       await run([
         collectRcJson,
         updateRcJson,
         askToSkip(
           'Setup user flow',
-          async (cfg: RcJson): Promise<RcJson> => {
-            const ufPath = cfg.collect.ufPath;
-            // DX create directory if it does ot exist
-            try {
-              readdirSync(ufPath);
-            } catch (e) {
-              mkdirSync(ufPath, {recursive: true});
-            }
-            addUserFlow(exampleName, ufPath);
-            return Promise.resolve(cfg);
-          },
+          generateUserFlow,
           {
             precondition: userflowIsNotCreated
           })
