+++ packages/cli/src/lib/commands/init/static/basic-navigation.ts
