+++ packages/cli/src/lib/commands/init/utils.ts
@@ -5,10 +5,7 @@ import { log } from '../../core/loggin';
 import { InitArgvOptions } from './options/types';
 import { CollectRcOptions, PersistRcOptions } from '../collect/options/types';
 import { AssertRcOptions } from '../assert/options/types';
-
-const FlowExampleMap: Record<FlowExamples, string> = {
-  'basic-navigation': 'order-coffee.uf.ts'
-};
+import { FlowExampleMap } from './constants';
 
 export function getExamplePathSrc(flowExample: FlowExamples, folder: string): string {
   const fileName = FlowExampleMap[flowExample];
@@ -18,22 +15,6 @@ export function getExamplePathSrc(flowExample: FlowExamples, folder: string): st
 export function getExamplePathDest(flowExample: FlowExamples, folder: string): string {
   const fileName = FlowExampleMap[flowExample];
   return join(folder, fileName);
-}
-
-export function addUserFlow(flowExample: FlowExamples, folder: string) {
-
-  const fileName = FlowExampleMap[flowExample];
-  const exampleSourceLocation = join(__dirname, 'static', fileName);
-  const exampleDestination = join(folder, fileName);
-
-  if (readFile(exampleDestination) !== '') {
-    throw new Error(`No flow example given for name ${flowExample}.`);
-  }
-
-  const fileContent = readFile(exampleSourceLocation, { fail: true }).toString();
-  writeFile(exampleDestination, fileContent);
-
-  log(`setup user-flow for basic navigation in ${folder} successfully`);
 }
 
 export function getInitCommandOptionsFromArgv(argv: any) {
